Fix windows build
Somehow there were some unfixed errors from enabled warnings?
I'm unsure if I've pulled repo in unstable state, or those were somehow
missed.

+++ Code/Editor/Include/Command.h
@@ -87,7 +87,7 @@ public:
                 return false;
             }
         }
-        int GetArgCount() const
+        size_t GetArgCount() const
         { return m_args.size(); }
         const AZStd::string& GetArg(int i) const
         {
